+++ packages/ledger-icp/src/ledger.canister.spec.ts
@@ -1,23 +1,18 @@
 import { ActorSubclass } from "@dfinity/agent";
-import { Memo, Payment, SendRequest } from "@dfinity/nns-proto";
 import { Principal } from "@dfinity/principal";
 import { arrayOfNumberToUint8Array } from "@dfinity/utils";
 import { mock } from "jest-mock-extended";
 import type { _SERVICE as LedgerService } from "../candid/ledger";
-import { AccountIdentifier } from "./account_identifier";
-import { toICPTs } from "./canisters/ledger/ledger.request.converts";
 import { TRANSACTION_FEE } from "./constants/constants";
 import {
   BadFeeError,
   InsufficientFundsError,
-  InvalidSenderError,
   TxCreatedInFutureError,
   TxDuplicateError,
   TxTooOldError,
 } from "./errors/ledger.errors";
 import { LedgerCanister } from "./ledger.canister";
 import { mockAccountIdentifier } from "./mocks/ledger.mock";
-import { E8s } from "./types/common";
 
 describe("LedgerCanister", () => {
   describe("accountBalance", () => {
@@ -87,40 +82,6 @@ describe("LedgerCanister", () => {
         const expectedFee = await ledger.transactionFee();
         expect(service.transfer_fee).toBeCalled();
         expect(expectedFee).toBe(fee);
-      });
-    });
-
-    describe("for hardware wallet", () => {
-      it("returns account balance with query call", async () => {
-        const queryFetcher = jest
-          .fn()
-          .mockResolvedValue(new Uint8Array(32).fill(0));
-        const ledger = LedgerCanister.create({
-          queryCallOverride: queryFetcher,
-          hardwareWallet: true,
-        });
-        const balance = await ledger.accountBalance({
-          accountIdentifier: mockAccountIdentifier,
-          certified: false,
-        });
-        expect(typeof balance).toEqual("bigint");
-        expect(queryFetcher).toBeCalled();
-      });
-
-      it("returns account balance with update call", async () => {
-        const updateFetcher = jest
-          .fn()
-          .mockResolvedValue(new Uint8Array(32).fill(0));
-        const ledger = LedgerCanister.create({
-          updateCallOverride: updateFetcher,
-          hardwareWallet: true,
-        });
-        const balance = await ledger.accountBalance({
-          accountIdentifier: mockAccountIdentifier,
-          certified: true,
-        });
-        expect(typeof balance).toEqual("bigint");
-        expect(updateFetcher).toBeCalled();
       });
     });
   });
@@ -401,237 +362,6 @@ describe("LedgerCanister", () => {
           });
 
         expect(call).rejects.toThrowError(TxCreatedInFutureError);
-      });
-    });
-
-    describe("for hardware wallet", () => {
-      const to = mockAccountIdentifier;
-      const amount = BigInt(100000);
-
-      it("handles invalid sender", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => {
-            throw new Error(`Reject code: 5
-                Reject text: Canister ryjl3-tyaaa-aaaaa-aaaba-cai trapped explicitly: Panicked at 'Sending from 2vxsx-fae is not allowed', rosetta-api/ledger_canister/src/main.rs:135:9`);
-          },
-          hardwareWallet: true,
-        });
-
-        const call = async () =>
-          await ledger.transfer({
-            to,
-            amount,
-          });
-
-        await expect(call).rejects.toThrow(new InvalidSenderError());
-      });
-
-      it("handles duplicate transaction", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => {
-            throw new Error(`Reject code: 5
-                Reject text: Canister ryjl3-tyaaa-aaaaa-aaaba-cai trapped explicitly: Panicked at 'transaction is a duplicate of another transaction in block 1235123', rosetta-api/ledger_canister/src/main.rs:135:9`);
-          },
-          hardwareWallet: true,
-        });
-
-        const call = async () =>
-          await ledger.transfer({
-            to,
-            amount,
-          });
-
-        await expect(call).rejects.toThrow(
-          new TxDuplicateError(BigInt(1235123)),
-        );
-      });
-
-      it("handles insufficient balance", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => {
-            throw new Error(`Reject code: 5
-                Reject text: Canister ryjl3-tyaaa-aaaaa-aaaba-cai trapped explicitly: Panicked at 'the debit account doesn't have enough funds to complete the transaction, current balance: 123.46789123', rosetta-api/ledger_canister/src/main.rs:135:9`);
-          },
-          hardwareWallet: true,
-        });
-
-        const call = async () =>
-          await ledger.transfer({
-            to,
-            amount,
-          });
-
-        await expect(call).rejects.toThrow(
-          new InsufficientFundsError(BigInt(12346789123)),
-        );
-      });
-
-      it("handles future tx", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => {
-            throw new Error(`Reject code: 5
-                Reject text: Canister ryjl3-tyaaa-aaaaa-aaaba-cai trapped explicitly: Panicked at 'transaction's created_at_time is in future', rosetta-api/ledger_canister/src/main.rs:135:9`);
-          },
-          hardwareWallet: true,
-        });
-
-        const call = async () =>
-          await ledger.transfer({
-            to,
-            amount,
-          });
-
-        await expect(call).rejects.toThrow(new TxCreatedInFutureError());
-      });
-
-      it("handles old tx", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => {
-            throw new Error(`Reject code: 5
-                Reject text: Canister ryjl3-tyaaa-aaaaa-aaaba-cai trapped explicitly: Panicked at 'transaction is older than 123 seconds', rosetta-api/ledger_canister/src/main.rs:135:9`);
-          },
-          hardwareWallet: true,
-        });
-
-        const call = async () =>
-          await ledger.transfer({
-            to,
-            amount,
-          });
-
-        await expect(call).rejects.toThrow(new TxTooOldError(123));
-      });
-
-      it("handles subaccount for hw", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: jest
-            .fn()
-            .mockResolvedValue(new Uint8Array(32).fill(0)),
-          hardwareWallet: true,
-        });
-        const fromSubAccount = [
-          0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-          0, 0, 0, 0, 0, 0, 0, 0, 1,
-        ];
-
-        const res = await ledger.transfer({
-          to,
-          amount,
-          fromSubAccount,
-        });
-
-        expect(typeof res).toEqual("bigint");
-      });
-
-      const initExpectedRequest = async ({
-        to,
-        amount,
-        memo,
-        fee,
-      }: {
-        to: AccountIdentifier;
-        amount: bigint;
-        memo?: bigint;
-        fee?: E8s;
-      }): Promise<SendRequest> => {
-        const expectedRequest = new SendRequest();
-        expectedRequest.setTo(await to.toProto());
-
-        const payment = new Payment();
-        payment.setReceiverGets(await toICPTs(amount));
-        expectedRequest.setPayment(payment);
-
-        const requestMemo = new Memo();
-        requestMemo.setMemo((memo ?? BigInt(0)).toString());
-        expectedRequest.setMemo(requestMemo);
-
-        expectedRequest.setMaxFee(await toICPTs(fee ?? TRANSACTION_FEE));
-
-        return expectedRequest;
-      };
-
-      it("should set a default fee for a transfer", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => Promise.resolve(new Uint8Array()),
-          hardwareWallet: true,
-        });
-
-        // @ts-ignore - private function
-        const spy = jest.spyOn(ledger, "updateFetcher");
-
-        const expectedRequest = await initExpectedRequest({ to, amount });
-
-        await ledger.transfer({
-          to,
-          amount,
-        });
-
-        expect(spy).toHaveBeenCalledWith({
-          // @ts-ignore - private variable
-          agent: ledger.agent,
-          // @ts-ignore - private variable
-          canisterId: ledger.canisterId,
-          methodName: "send_pb",
-          arg: expectedRequest.serializeBinary(),
-        });
-      });
-
-      it("should use custom fee for a transfer", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => Promise.resolve(new Uint8Array()),
-          hardwareWallet: true,
-        });
-
-        // @ts-ignore - private function
-        const spy = jest.spyOn(ledger, "updateFetcher");
-
-        const fee = BigInt(990_000);
-
-        const expectedRequest = await initExpectedRequest({ to, amount, fee });
-
-        await ledger.transfer({
-          to,
-          amount,
-          fee,
-        });
-
-        expect(spy).toHaveBeenCalledWith({
-          // @ts-ignore - private variable
-          agent: ledger.agent,
-          // @ts-ignore - private variable
-          canisterId: ledger.canisterId,
-          methodName: "send_pb",
-          arg: expectedRequest.serializeBinary(),
-        });
-      });
-
-      it("should use custom memo for a transfer", async () => {
-        const ledger = LedgerCanister.create({
-          updateCallOverride: () => Promise.resolve(new Uint8Array()),
-          hardwareWallet: true,
-        });
-
-        // @ts-ignore - private function
-        const spy = jest.spyOn(ledger, "updateFetcher");
-
-        const memo = BigInt(990_000);
-
-        const expectedRequest = await initExpectedRequest({ to, amount, memo });
-
-        await ledger.transfer({
-          to,
-          amount,
-          memo,
-        });
-
-        expect(spy).toHaveBeenCalledWith({
-          // @ts-ignore - private variable
-          agent: ledger.agent,
-          // @ts-ignore - private variable
-          canisterId: ledger.canisterId,
-          methodName: "send_pb",
-          arg: expectedRequest.serializeBinary(),
-        });
       });
     });
   });
@@ -929,76 +659,6 @@ describe("LedgerCanister", () => {
           });
 
         expect(call).rejects.toThrowError(TxCreatedInFutureError);
-      });
-    });
-
-    describe("for hardware wallet", () => {
-      const to = {
-        owner: Principal.fromHex("abcd"),
-        subaccount: [] as [],
-      };
-      const amount = BigInt(100000);
-
-      it("should set a default fee for a transfer", async () => {
-        const service = mock<ActorSubclass<LedgerService>>();
-        service.icrc1_transfer.mockResolvedValue({
-          Ok: BigInt(1234),
-        });
-        const ledger = LedgerCanister.create({
-          certifiedServiceOverride: service,
-          serviceOverride: service,
-          hardwareWallet: true,
-        });
-
-        const icrc1Memo = new Uint8Array();
-        await ledger.icrc1Transfer({
-          to,
-          amount,
-          icrc1Memo,
-        });
-
-        expect(service.transfer_fee).not.toBeCalled();
-
-        expect(service.icrc1_transfer).toBeCalledWith({
-          to,
-          fee: [BigInt(10000)],
-          amount,
-          memo: [icrc1Memo],
-          created_at_time: [],
-          from_subaccount: [],
-        });
-      });
-
-      it("should use custom fee for a transfer", async () => {
-        const service = mock<ActorSubclass<LedgerService>>();
-        service.icrc1_transfer.mockResolvedValue({
-          Ok: BigInt(1234),
-        });
-        const ledger = LedgerCanister.create({
-          certifiedServiceOverride: service,
-          serviceOverride: service,
-          hardwareWallet: true,
-        });
-
-        const fee = BigInt(990_000);
-        const icrc1Memo = new Uint8Array();
-        await ledger.icrc1Transfer({
-          to,
-          amount,
-          fee,
-          icrc1Memo,
-        });
-
-        expect(service.transfer_fee).not.toBeCalled();
-
-        expect(service.icrc1_transfer).toBeCalledWith({
-          to,
-          fee: [fee],
-          amount,
-          memo: [icrc1Memo],
-          created_at_time: [],
-          from_subaccount: [],
-        });
       });
     });
   });